Mods to the migration

+++ database/migrations/2017_04_17_080333_create_map_locations_table.php
@@ -17,8 +17,8 @@ class CreateMapLocationsTable extends Migration
             $table->increments('id');
             $table->integer('company_id')->unsigned();
             $table->foreign('company_id')->references('id')->on('companies');
-            $table->float('lng');
-            $table->float('lat');
+            $table->double('lng');
+            $table->double('lat');
             $table->timestamps();
         });
     }
